feat: enhance DELETE route for tester entity with error handling and authentication

+++ src/api/tester/[testerId]/route.ts
@@ -1,5 +1,6 @@
 import { MedusaRequest, MedusaResponse } from "@medusajs/framework";
 import { TESTER_MODULE } from "../../../modules/tester";
+import { MedusaError, Modules } from "@medusajs/framework/utils";
 
 export async function DELETE(req: MedusaRequest, res: MedusaResponse) {
 	const { testerId } = req.params;
@@ -9,6 +10,28 @@ export async function DELETE(req: MedusaRequest, res: MedusaResponse) {
 	}
 
 	const testerModuleService = req.scope.resolve(TESTER_MODULE);
+	const authModuleService = req.scope.resolve(Modules.AUTH);
+	const tester = await testerModuleService.retrieveTester(testerId);
+	if (!tester) {
+		throw new MedusaError(
+			MedusaError.Types.NOT_FOUND,
+			"Tester with the given ID does not exist"
+		);
+	}
+
+	const providerIdentity = await authModuleService.listProviderIdentities({
+		entity_id: tester.email,
+	});
+	if (!providerIdentity.length) {
+		throw new MedusaError(
+			MedusaError.Types.NOT_FOUND,
+			"Tester not found in provider identities"
+		);
+	}
+	await authModuleService.deleteProviderIdentities([providerIdentity[0].id]);
+	await authModuleService.deleteAuthIdentities([
+		providerIdentity[0].auth_identity_id!,
+	]);
 	await testerModuleService.deleteTesters(testerId);
 
 	return res.status(204).send();
